requirements.txt
fastapi
uvicorn
opencv-python-headless
mediapipe
numpy
pillow
deepface
openai
python-multipart
